[bugfix] Throw an exception on all processes if the grid is too small
If overlapping computation on a too small grid is requested, an exception
is thrown by the YaspGrid constructors. However, the condition that is
checked for that does not trigger on all processes. The processes where no
exception was thrown did not terminate. This commit solves the problem
by communicating the result of the check to all processes.