[vtkfunction][regression][release] Generalize P1VTKFunction to work also with std::vector
Commit 568fdf7a90c3dd0ac874016cfae94ad82d04aaf9 broke the ability
of P1VTKFunction to handle data stored in a V=std::vector since it
requires V::field_type, which is not offered by std::vector. This
patch uses ctype instead which come closest to the situation before
the offending commit.

Jö> This fixes a regression, that's why it should go into the release.  See
Jö> FS#1700.