[Rangegenerators][Release] The default entities() method should not enforce Partitions::all
When I originally wrote the range generator functions, I implemented the
entities() function that does not specify a parallel partition as a
forward to the function with a parallel partition and hardcoded the
partition to Partitions::all. This turned out to be a *really* bad idea
because it makes it breaks custom GridView-like objects over smaller
partitions in generic code that does not want to care about the
partition.

I know this commit is late, but is absolutely crucial for PDELab's
nonoverlapping parallel support. Carsten, would you please consider
taking this one for the release?