[tests][vtk][#flyspray/FS#1652] Complain about missing python vtk only once.
The test function checkVTKFile() would check once whether the python vtk
module was available and cache the result of the check for later invocations.
However, it would still print the warning about missing python vtk on every
invocation.  Change it to print the warning only once.

Also add a comment that explains that the check is indeed only done once, for
the weary programmer who isn't closely familiar with the semantics of
initializing static variables.