[UG] Fix building UG on GCC 4.4
GCC 4.4 has a weird bug where it is too stupid to compare prototypes of
declared methods within the class body with the actual implementation
outside of it if a little SFINAE is involved.

Fixed the brute force way by moving the implementations inside the class
body. Yay for supporting broken stone-age compilers ^^