[dgf][yaspgrid] Throw more meaningful exception, if file is not found
If the DGF file could not be found in the DGFGridFactory for YaspGrid,
you got the exception 'file not found in BasicBlock::BasicBlock'. As the
BasicBlock only gets an istream, it cannot display the filename.
Therefore, we directly check whether the ifstream is valid before
passing it anywhere. If not, we through an exception containing the
file name.