[YaspGrid] Make YGridComponent iterators bidirectional
This patch adds an `operator--()` to the `YGridComponent` iterator. In
order to simplify detecting invalid (out-of-range) iterators, the
iterator now carries a sentinel flag that is true for sentinel
iterators, which also reduces the cost of comparing valid and invalid
iterators, as the comparison stops after checking the sentinel flag.

Finally, there are now methods `rbegin()` and `rend()` for iterating in
reverse direction. Note, however, that this returns the same type of
iterator as `begin()` and `end()` and you thus have to use
`operator--()` to iterate in reverse direction.