[YaspGrid] Make YGrid iterator bidirectional
Building on top of the bidirectional `YGridComponent` iterator, this
patch also makes the `YGrid` iterator bidirectional.

Importantly, this requires changing the `_which` member of the iterator,
which is the index of the current `YGridComponent`, to a signed value,
allowing it to represent an `rend()` iterator as -1. Moreover, the
`YGrid` now also contains pointers to the ends and before the begins of
the contained `YGridComponent`s.

As with the `YGridComponent`, there are methods `rbegin()` and `rend()`
for reverse iteration, and reverse iteration has to occur using
`operator--()`.