Factor out the type for process ranks into an explicitly given type 'Rank'
This type is hard-wired to 'unsigned int'.  Originally, I wanted it to be
whatever comm().rank() returns, i.e., the type actually used by MPI for
process ranks.  However, that is an interface change, because the UGGrid
and ParMetisPartitioner interface contain std::vectors of the rank type.
And these are not converted automatically.