Introduce method 'fixedSize', which does the same as 'fixedsize', but has camel-case naming
The CommDataHandleIF interface class has a member 'fixedsize'.  This is one
of the few in dune-grid that does not comply with the official naming rules:
the name should be 'fixedSize'.  This patch introduces a new method 'fixedSize',
to allow a gradual transition from the old spelling to the new one.
The make the transition as painless as possible, but methods implement the
following behavior (remember that CommDataHandleIF is the base class in a
CRTP pattern):

- If the derived/implementation class has a method 'fixedSize', then that
  method is called.
- Otherwise, the method 'fixedsize' is called.

This approach allows both implementors of grids and implementors of data
handles to switch their code from 'fixedsize' to 'fixedSize' at their
convenience.

Finally, I have come across a usecase for CRTP. :-)