Simplification: merge the two classes that implement element facets in 3d into one
Previously, the implementation of UGGrid entities for facets of 3d elements
consisted of two classes: UGGridEntity<1,3>, a specialization of the generic
implementation class for UGGrid entities, and UGFaceEntity, a special
implementation class for the codim-1 case in 3d.  There was never a real
reason for this separation into two classes, and this patch does away
with it.  This leads to a minor code simplification.

Note: The same separation into two classes holds for grid edges, and there is
a reason.  The UGEdgeEntity class, which contains the actual implementation,
serves for *two* cases: UGGridEntity<1,2> and UGGridEntity<2,3>.
C++ partial specialization rules apparently do not allow to merge all of this
into one class.