[bugfix] make subIndex method for OneDGrid compile for all codimension
The index sets of OneDGrid support the subIndex method for entities of
all codimensions. However, while the code was semantically correct for
the case "cc != 0", the compiler would try to compile the "cc == 0"
branch, too. This caused a compile-time error.

This patch fixes the issue by calling a separate method for each
codimension.