[test] add static check for higher codimension information on grid views
Until now, the static part of the grid check only check on codim 0 structures,
e.g., on the index set. This patch also performs these checks on higher
codimensions.

For the IndexSet::subIndex method, it will only check whether the method
compiles. If codim > 0, it will silently ignore a Dune::NotImplemented
exception, so that grids are not forced to implement the method; it simply has
to compile.

Note: This check is able to detect the error fix in !86.