Implement the subIndex method for entities of all dimensions
Previously, it was available only for elements.

The following question appears: what does the codim-parameter in

  template<int cc>
  unsigned int subIndex (const Codim<cc>::Entity& e,
                         int i,
                         unsigned int codim) const

mean?  Is it the codimension wrt to the grid, or wrt to the entity e?
I did not find this documented anywhere, and therefore made my own
choice.  The 'codim' in this patch is the codimension with respect
to the *grid*.