[doc] document sementics of codim parameter for IndexSet::subIndex
If an entity of codimension cc > 0 is passed into IndexSet::subIndex, what
is the semantics of the 'codim' parameter? In DUNE, codimensions are always
with respect to the grid dimension (see method subEntity of reference element).

This patch clearly states that codim is to be understood with respect to the
grid dimension.