Merge branch 'feature/fix-uggrid-subindex' into 'master'
Fix subIndex method for UGGrid

When using Hygrid::ifElse always keep in mind that some types in
the functions given as "if" and "else" branch are fixed, such that
compilation may fail altough the template is not instanciated.

To avoid this, those functions have one templated parameter which
is instanciated with the identity for the selected branch. Passing
objects through this branch will delay resolution of their type
until instantiation which does only happen for the selected branch.

By passing the entity through `id` we use this to make subIndex
compile for clang. It seems that gcc is less picky here.

See merge request !101