Do not force grid to use Dune::Entity.
This is test code and should be as generic as it can be.
Previously (but not since ever) it assumed that Grid::Entity
is always a subclass of Dune::Entity. This seems to hard a
restriction as it is only used to check that the grid has
a hasEntity<codim> capability for all its Entities (which is a
bug IMHO, too).

Therefore this commit makes CheckSubEntityGeometry a template struct
that has the Grid as a parameter and a nested Operation struct with
template parameter codim that is used in the ForLoop for testing.