Merge branch 'do-not-force-grids-to-subclass-master' into 'master'
Do not force grids to subclass Entity or Geometry

The current implementation assumes that both the Entity and the Geometry of a grid are subclasses of Dune::Entity and Dune::Geometry. This requirement must have been introduced after 2.4.1 and is IMHO to much of restriction. After all these are just tests. With this MR we drop this additional restriction. Tested with opm-grid.

Closes #36 

Should be in the release.

See merge request !114