Increase timeouts for VTKWriter tests [ci skip]
The VTK tests are close to the timeout on the CI system and sometimes
fail.

In addition the parallel vtktest has one process busy with polling in a
MPI function, taking 100% CPU for doing nothing, while the main process
is running the Python script. In case of parallel builds, this
unexpected resource usage slows down the test even more.