Append to-be-visited sons at the end of the list
Prepending them to the beginning of the list would mean they are never
visited as the iterator only moves forward.  This required switching to
a `std::list` as the `std::forward_list` does not provide methods to add
items at the end.

This regression was introduced in 20b821be632f5c8c1a30b1b964216402357bfafc
which replaced `Dune::SLList` with `std::forward_list`, but changed
`push_back` to `push_front` as well.

Closes: #53