Add `update()` method to `CommDataHandleIF`
When using a grid's load balancing feature, the following steps
happen:

1. User data is gathered (`gather`)
2. The grid is manipulated
3. User data is scattered again (`scatter`)

User data is often stored in a vector and accessed via the entity
indices obtained from a mapper.  However between steps (2.) and (3.)
the grid is manipulated and the vector size and mappers need to be
updated before they can be used to store the scattered user data.

This patch adds a `update()` method that must be called by grid
implementations between step (2.) and (3.) to allow updating any
vectors and mappers used to store data.