[geometry] Add referenceElement() function for grid geometries
This patch adds a function `referenceElement(Geometry)` that returns
the reference element for a grid geometry. In order to allow grid
implementors to return custom reference elements (e.g. for polyhedra),
it performs a second-level lookup by calling `referenceElement()` with
both the geometry facade and the geometry implementation stored inside
the facade class.

The default implementation of the second-level `referenceElement()`
function just looks up the default reference element for the
GeometryType of the geometry. Grid implementations that want different
behavior should overload this function for their specific geometry
implementation(s).