[referenceelement] Make referenceElement(Geometry) work for the broken GCC 4.9 of Debian 8
GCC 4.9 on Debian 8 trips over the `friend auto referenceElement(...)`
in Geometry. I can't really debug that as I don't have Debian and my
Homebrew GCC 4.9 compiles the code just fine. So this is a brute-force
workaround that sidesteps the problem explicitely for GCC < 5 (I hope we can
abandon that antiquity sooner rather than later anyway).

The hack is rather ugly, as I couldn't find a way to make
`referenceElement()` a friend of Geometry (the compilers tripped about
all kinds of other problems when adding a trailing return type), so we
introduce a helper struct that takes over friend duties.