SCSGMapper: Make 'size' return the same type as 'IndexSet::size'
Previously, the SCSGMapper class had 'int' hard-coded as the
return type of the 'size' method.  This method changes it
to whatever the underlying IndexSet is using for its 'size'
method.  I think this should settle the infamous
'signed-vs.-unsigned' discussion at least for this method.

Additionally, the new type is exported as 'size_type'.