Rudimentary implementation of 'id' method for facets (codim1 entities)
This patch implements the 'id' method for entities of codimension 1
in a three-dimensional UGGrid.  The implementation is incomplete:
it will work in general, but it will return wrong results if a facet
has a copy on a coarser level.

Nevertheless, the current implementation makes the test pass again.
I will improve the implementation (and sharpen the test) in a
subsequent patch.