Non-MPI: encode codimension into entity id
When UGGrid is configured with MPI, then the ids produced by
dune-uggrid are injective across all codimensions and
GeometryType.  This is the way it should be in Dune (modulo
the 'copy' rule).  However, when MPI is not used, then two
entities of different codimension can have the same id.
To fix this, we now encode the entity codimension into the
highest two bits of the id.  This was already done in the
code previously, but not for all codimensions, and not in
a very legible way.