Make method MessageBufferIF::read non-const
The method CommDataHandleIF::scatter receives a non-const
reference to the message buffer.  This reference is
non-const because reading from the buffer advances the
internal iterator, and hence changes the buffer state.
However, the MessageBufferIF claims that the read method
is const.  This patch removes the const to make the
interface consistent.