Allow MCMGMapper to allocate multiple indices per geometry type
This patch modifies MCMGMapper such that multiple indices may be allocated
per geometry type. To this end, the layout function does no longer return a
bool but the number of indices to allocate for a given geometry type.
Old layout functions continue to work, because `bool` implicitly casts into
`std::size_t` such that `true` is 1 and `false` is 0 so that exactly one
index is allocated, if the geometry type was included in the layout.

Semantic enhancements:
- The `index` method returns the first index allocated for a geometry type
- The `contains` method returns whether at least one index was allocated
  for a geometry type (and fills in the first index).
- a new method `indices` returns an IntegralRange containing all indices
  allocated for a given geometry type. It can be used in a range-based-for
  like this:
  ```
  for( auto i : mapper.indices( element ) )
    doSomething( i );
  ```