Rewrite the documentation for the IdSet class
I wanted to state a few assumptions on the IdType that have been
implicit all these years, like copy-assignability of and default
constructibility.  I ended up rewriting most of the class
documentation of the IdSet class, which hadn't been touched in
a long time.  I hope my changes are for the better.