[vtkwriter] detect whether `bind`/`unbind` are present instead of `localContext`
Currently the VTKWriter detects whether a function provides a method
`localContext` to find out whether it may call `bind` and `unbind`.
It will never actually call `localContext`.

This behavior is quite surprising. To fix the issue, this patch actually
checks for the methods that are actually called, i.e., `bind` and `unbind`.