uggrid: get rid of `UGMakeableEntity`
Just use `Entity` directly instead and access the `setToTarget()`
method via `entity.impl().setToTarget()`.

With this chage the user-provided `gather` and `scatter` methods are
also called with instances of the `Entity` facade class instead of a
derived class.  This makes it easier to overload them for different
codimensions of entities (by overloading them for the different types
of `Entity`).