[bugfix] Do not rely on `operator[]` on a `JacobianInverseTransposed`
Some grid implementations, like dune-spgrid, will not return a dense
matrix for the `JacobianInverseTransposed`. While we could simply cast it
into a `FieldMatrix`, this might reduce performance due to unnecessary
dense matrix vector multiplications.

This patch uses the `mv` method, which must be provided by any
`JacobianInverseTransposed` instead. Apart from retaining the
performance of the specialized implementation, this improves the overall
clarity of the code.

As a drawback, the local and global gradient vectors are copied for the
time being. This unnessary copy can be avoided by a specialized
`DenseVector` implementation for a single axis in a pybind11::array
accessor.