Do not throw when asked for negative codimensions
Currently, an exception is thrown when OneDGrid is asked
for the number of entities on a given level with a given
negative codimension.  AFAIK this behavior is not actually
required by any formal specification.  As the LagrangeBasis
class from dune-functions currently does such calls
(only to disregard the result later) I am changing OneDGrid
such that the size(level,codim) method returns 0 for
negative codimensions.