Make Geometry export the return type of the 'volume' method
Previously this was 'ctype', but grids may care about proper
dimensions.  In that case, a volume is not the same as a
coordinate, and a separate type for volumes is necessary.

The interface class does not hard-wire any type but rather exports
the type used by the Geometry implementation class.  That type
is determined by introspection, there is no need to adapt
downstream grid implementations.