Get rid of the UGGrid::heapSize_ data member
The hand-coded memory heap of UG has been removed from dune-uggrid
about a year ago.  Hence, the heapSize parameter of UGGrid does
not have any purpose anymore.  This patch removes it completely.
It only leaves the public method setDefaultHeapSize and marks it
as deprecated, in order not to break people's codes.