GmshReader: deprecate method that can't be used correctly
GmshReader::read(fileName, boundarySegmentToPhysicalEntity,
                 elementToPhysicalEntity, verbose, bool)

can't be used correctly, as the vector of elementToPhysicalEntity is indexed
by the insertion index.  But there is no way to find out the insertion index,
as you need a grid factory to do that.  And there are better alternatives,
either mthods that don't take data vectors as arguments, or ones that also
pass along a gridfactory.

Addresses: flyspray/FS#1698