GmshReader: deprecate optional arg that is easy to use incorrectly
In
```c++
GmshReader::read(factory, fileName, boundarySegmentToPhysicalEntity,
                 elementToPhysicalEntity, verbose = true,
                 insertBoundarySegments = true)
```

the default argument insertBoundarySegments is easy to use in the wrong way.
If set to `false`, the data read into the `boundarySegmentToPhysicalEntity`
vector cant be used, as the indexing into that vector happens according to the
insertionIndex of the boundary segments, but those were not inserted.

Addresses: flyspray/FS#1698