GmshReader: more flexible and robust io-arg-style interface for reading data
This allows the user to not pass the data vectors he/she is not interested in.
To ignore a vector, pass `std::ignore`.

In addition, in place of the boundary segment data a `bool` value may be
passed, stating whether or not to insert boundary segments.  `false` is
equivalent to `std::ignore` and no boundary segments are inserted.  When a
vector to fill is given, boundary segments are always inserted (otherwise the
user would be unable to correctly index that vector).