Help compiler to show more useful errors
Concepts do not allow recursion and hence, we have our recursive algorithm moved
into specializated templated integral constants. Unfortunately, when integral constants evaluate
to false, the compiler has no chance to inspect the underlying concepts and show a proper backtrace.
Now, in this commit, we just add a concept check on the codimension 0 because an error in any
codimension is very likely to show in others. This way users wil get better error messages if
the error happens for all codimensions. If the interface does not matche in other codimensions an
error is still produced without the concepts backtrace.