Simplify the UGGridEntity implementation
The implementation of UGGridEntity objects used specializations
for the different codimensions.  This lead to lots of redundant code.
Modern C++17, in particular 'if constexpr' allows to get rid
of most of that redundancy.  The new implementation has only
two classes:  One for the elements and one for all the rest.