Bugfix: communicate only on toplevel leaf edge entities
The implementation of the isLeaf method in UGGrid will return true
for actual leaf edges and (some of) its lower-level copies.
This is intentional, but it lead the UGGrid subdomain communication
code to sometimes call gather and scatter too many times on
such leaf entities.  This patch introduces a custom isLeaf method
that really only marks the top-most copies of leaf edges.

TODO: Also fix this for facet entities in 3d grids.