UGGrid: Test whether gather/scatter are called correctly
By correctly we mean: The appropriate number of times for each
entity, given the entity's partition type and its partition type
on other processes.  Testing this uncovered a surprising number
of bugs when using the InteriorBorder_All_Interface.  All of
them are fixed now, but I expect more when testing the other
interfaces.

To implement this test, the UGGridEntity class now has a
non-interface method 'partitionTypes' that, for a distributed entity,
returns the partition type on each process that stores the entity.