Bugfix in UGGridIdSet::id method for facets of 3d elements
The Dune grid interface requires that copies of entities on different
levels have the same id.  The UGGridIdSet failed to implement this
in the 'id' method for codim-1 faces of 3d elements.  I am afraid
I omitted that out of pure laziness back in the days when I wrote
that code.

The commit also adds a simply unit test that would have caught
the bug.