[YaspGrid] Do not precompute intersection geometries
This simplifies the code a bit.  I have not measured the performance
impact, but I believe that the new code will be faster if its speed
changes at all.  YaspGrid intersection geometries are simple objects,
and constructing them on demand allows for more aggressive optimization.

As a side effect, this change makes the default constructor of
YaspGridGeometry unnecessary.  This is what motived this patch,
see https://gitlab.dune-project.org/core/dune-geometry/-/issues/25
for some information.