Add method update(gridView) method to (SCSG|MCMG)Mapper
Since there's no guarantee that a `GridView` or `IndexSet&`
remains valid after grid adaptation, they must be updated
inside of the mappers. For the `MCMGMapper` the only way
to achieve this was
```cpp
  mapper = MCMGMapper(gridView, mapper.layout());
```
For the `SCSGMapper` it wasn't possible at all, because
it stores a reference to the `IndexSet`. This patch
adds new interface methods `mapper.update(gridView)` to allows
a clean update.

Notice that the `Mapper` interface class is not modified,
because it does not know about grid views (and is outdated
anyway).