Test with several UGGrids at once even when MPI is enabled
For a long time, UG would only allow to have one grid at
a time when it was built with MPI enabled.  Therefore,
the test-ug test program had to skip several tests in
that case.  However, that UG shortcoming has been fixed
several years ago, and it is now possible to have more
than one grid object even with MPI.  Therefore, the need
to skip tests is no more, and the corresponding code
can be removed.