Reduce the number of undef statements in the ug_undefs.hh
Many of those undef statements refer to macros that have been removed from the dune-uggrid code.  The number of macros is reduced from 1633 to 1162.

It is required to include the `ugincludes.hh` file multiple times with different macros defined. This allows to have 2d and 3d code in one library. But, this requires to undefine the include-guard if this file. Previously this was done inside the `ug_undefs.hh` what makes it complicated to automatically cleanup this file. Thus, this `#undef` is moved to the place where the `ugincludes.hh` file is actually included.