Add jacobian()... to GeometryReference
This adds `jacobian()`, `jacobianInverse()`, `Jacobain`, and `JacobianInverse`
to the `Dune::GeometryReference` class. The implementation works exactly
as in `Dune::Geometry`: If the method and typedef is provided by the
implementation, then this one is used. Otherwise the default version
with `Dune::transpose()` drops in.

This is needed because `Dune::GeometryReference` is - altough it's
contained in `albertagrid/` - in fact a generic interface class that
is e.g. also used in dune-spgrid.