[test]Add check if entities are assignable
There is no canonical place where entities are checked.
`checkEntityLifetimeForCodim()` does some entity checks, but
only for codims, where there's an iterator.
Vy putting it into the `subIndexCheck()` it's also tested
if this is not the case and entities are only available via
`subEntities(e, cd)`.

Currently this makes the gridcheck fail for `UGGrid`.