Rewrite a few iterator loops with range-based 'for'
The 'codespell' tool complained about the variable name 'wit'
that is used a few times as a loop variable.  Rather than
having codespell ignore it this patch simply rewrites the
old iterator loops using range-based 'for'.  That simplifies
the code and gets rid of the codespell error at the same time.