Simplify UGGridGeometry::corner using constexpr if
The UGGrid code dates to before constexpr if existed, and uses
particular "pointless" casts to achieve a similar effect.
Replace some of these casts by constexpr if, which is much easier
to read.

The code is simplified further because UGGrid has recently started
to use FieldVector for point coordinates.  Therefore, componentwise
copying is not necessary anymore.