[UGGrid] Simplify computation of integrationElement
UGGrid has used an overly complicated way to compute the
integrationElement since practically forever.  Instead of computing
the determinant of the Jacobian of the geometry transformation,
it computed the inverse determinant of the inverse Jacobian.
This lead to a matrix inversion being performed at each call
that wasn't actually necessary.

This patch removes this pointless inversion.