Use std::variant in intersection to store a face, a face vector, or none
The common case of an intersection is when it has only one face. For this case we use a variant that either holds one face, a vector of faces, or none. This way the common case is stored in the stack whereas the general case is still possible but needs dynamic allocation