Add `HAVE_STDBOOL_H` define for Alberta when `stdbool.h` is available
In older versions of Alberta, it tries to fix the issues around not having
proper bool in C by redefining bool as _Bool or vice versa, depending
on the language that it is compiled with (C or C++). Clang's stdbool header
does essentially the same and we end up with `typedef bool bool` which is a
compile error. This whole bool/_Bool logic is inside an ifdef in
Albert that is active if HAVE_STDBOOL_H is not defined, in which case
it just includes stdbool.h and uses that. So we can check during
CMake config if that header is available and set the flag accordingly.