UGGrid: Remove the rudimentary saveState/loadState implementations
The implementations of these methods looked like they do something
reasonable, but in fact they don't: They do call corresponding
UGGrid methods, but those methods never worked (in Dune!).

After giving it some thought it is now my opinion that the UGGrid
saveState/loadState functionality is not worth the time needed
to understand it and get it to work.  A modern snapshotting code
should be written from scratch in any case, using existing modern
standards as much as possible.

Therefore, this patch removes the dummy calls to SaveMultiGrid
and LoadMultiGrid.