Allow grid concept to handle missing entity types
The grid allows to have some entities of the grid missing for specific codimensions. However, the current concept check requires that all entities are present. This this is fixed by checking that the grid has this capability, and for index sets and entities, if the Codim<cc> traits are available.

Since a general entity alone cannot tell whether a given subentity should exist or not (due to the capabilities), they only check them if they exist. Therefore, this patch also adds an extra consistency check in the grid where we make sure that subentities exist if the capability is enabled.