Move the GmshReaderParser class into a separate file gmsh2parser.hh
At the same time, rename GmshReaderParser to Gmsh2Parser, and move it
into the `Dune::Impl::Gmsh` namespace. Since this is technically an
interface change (because the old class was not in the Impl namespace),
this commit also introduces a type alias with the old name.
The alias is deprecated right away.