Import some code from the dune-gmsh4 module
This is the code to do basic reading of Version 4 Ascii and binary Gmsh
files. More fancy stuff the like GridGenerators are left out.  There is
also a bit of reformatting and we are adding SPDX headers.

We also place all new code into the Impl::Gmsh namespace. This include
the Gmsh4Reader class, which is the central public interface method in the
dune-gmsh4 module. This is to postpone the discussion about interface
changes to later.

Simon Praetorius, who wrote the new code, licensed into under the
LGPL-3.0-or-later license in addition to the standard Dune license. This
dual licensing in kept in this code import.